started working on the new design

+++ branches/work/frame/src/objectselector.cpp
@@ -73,7 +73,7 @@ bool  ObjectSelector::full()const{
 	return d.sz == d.sv.size();
 }
 
-bool ObjectSelector::init(ulong _cp){
+bool ObjectSelector::init(size_t _cp){
 	d.sv.resize(fast_padding_size(_cp, 2));
 	setCurrentTimeSpecific(d.ctimepos);
 	//objq.reserve(_cp);
@@ -89,7 +89,7 @@ bool ObjectSelector::init(ulong _cp){
 void ObjectSelector::prepare(){
 	setCurrentTimeSpecific(d.ctimepos);
 }
-void ObjectSelector::raise(uint32 _pos){
+void ObjectSelector::raise(size_t _pos){
 	Locker<Mutex> lock(d.mtx);
 	d.uiq.push(_pos);
 	d.cnd.signal();
